+++ src/main/java/gregtech/common/redstonecircuits/GT_Circuit_Pulser.java
@@ -49,8 +49,8 @@ public class GT_Circuit_Pulser extends GT_CircuitryBehavior {
             }
         }
         aRedstoneCircuitBlock.setRedstone(
-                (byte) ((aCircuitData[4] > 0) && (aCircuitData[4] <= aCircuitData[0]) ? (byte) aCircuitData[1]
-                        : (aCircuitData[1] <= 0) || (aCircuitData[1] > 15) ? (byte) aCircuitData[5] : 0),
+                (aCircuitData[4] > 0) && (aCircuitData[4] <= aCircuitData[0]) ? (byte) aCircuitData[1]
+                        : (aCircuitData[1] <= 0) || (aCircuitData[1] > 15) ? (byte) aCircuitData[5] : 0,
                 aRedstoneCircuitBlock.getOutputFacing());
     }
